Changed the behavior of summary_and_sections strategy to match pipeline behavior
